+++ lib/src/specs/annotation.dart
@@ -10,29 +10,47 @@ import 'package:meta/meta.dart';
 import '../base.dart';
 import '../visitors.dart';
 import 'code.dart';
+import 'expression.dart';
 
 part 'annotation.g.dart';
 
+@Deprecated('Use an Expression node instead. Will be removed in 3.0.0.')
 @immutable
-abstract class Annotation
+abstract class Annotation extends Expression
     implements Built<Annotation, AnnotationBuilder>, Spec {
   factory Annotation([void updates(AnnotationBuilder b)]) = _$Annotation;
 
   Annotation._();
 
   /// Part after the `@` in annotation.
-  Code get code;
+  @override
+  Code get code => codeForAnnotation;
+
+  /// Use [code] instead.
+  ///
+  /// This property exists in order to keep compatibility with having a concrete
+  /// [Annotation] class, but also allow migrating to using [Expression] instead
+  /// to add annotations.
+  @protected
+  Code get codeForAnnotation;
 
   @override
-  R accept<R>(SpecVisitor<R> visitor, [R context]) =>
-      visitor.visitAnnotation(this, context);
+  R accept<R>(SpecVisitor<R> visitor, [R context]) {
+    return visitor.visitAnnotation(this, context);
+  }
 }
 
 abstract class AnnotationBuilder
-    implements Builder<Annotation, AnnotationBuilder> {
+    // ignore: deprecated_member_use
+    implements
+        Builder<Annotation, AnnotationBuilder> {
   factory AnnotationBuilder() = _$AnnotationBuilder;
 
   AnnotationBuilder._();
 
-  Code code;
+  Code get code => codeForAnnotation;
+  set code(Code code) => codeForAnnotation = code;
+
+  @protected
+  Code codeForAnnotation;
 }