+++ lib/src/specs/expression.dart
@@ -173,7 +173,8 @@ abstract class Expression implements Spec {
   }
 
   /// Returns an annotation as a result of calling this constructor.
-  Annotation annotation([
+  @Deprecated('Use "call" instead. Will be removed in 3.0.0.')
+  Expression annotation([
     List<Expression> positionalArguments,
     Map<String, Expression> namedArguments = const {},
     List<Reference> typeArguments = const [],
@@ -195,12 +196,14 @@ abstract class Expression implements Spec {
   }
 
   /// Returns an annotation as a result of calling a named constructor.
-  Annotation annotationNamed(
+  @Deprecated('Use a combination of "property" and "call". Removing in 3.0.0.')
+  Expression annotationNamed(
     String name,
     List<Expression> positionalArguments, [
     Map<String, Expression> namedArguments = const {},
     List<Reference> typeArguments = const [],
   ]) {
+    // ignore: deprecated_member_use
     return new Annotation((b) => b
       ..code = new InvokeExpression._(
               this, positionalArguments, namedArguments, typeArguments, name)