+++ test/e2e/injection_test.dart
@@ -35,8 +35,7 @@ void main() {
         ]).code
         ..lambda = true
         ..returns = $Thing
-        ..annotations
-            .add(new Annotation((b) => b..code = const Code('override')))));
+        ..annotations.add(refer('override'))));
 
     expect(
       clazz.build(),